Refactor frontend toward FSD boundaries
