+++ src/entities/agent/index.ts
@@ -1,0 +1,1 @@
+export type { AgentDescriptor } from "./model";
