+++ src/features/agent-run/index.ts
@@ -1,0 +1,10 @@
+export { closeWorkbenchTab } from "./tabActions";
+export { installAgentRuntime } from "./runtime";
+export { useAgentRun } from "./useAgentRun";
+export {
+  createTabState,
+  selectTab,
+  useWorkbenchStore,
+  type FollowUpQueueItem,
+  type TabState,
+} from "./model";
