+++ src/features/goal-input/api.ts
@@ -1,0 +1,5 @@
+import { invokeCommand } from "../../shared/api";
+
+export function loadGoalFile(path: string) {
+  return invokeCommand<string>("load_goal_file", { path });
+}
