+++ src/features/goal-input/index.ts
@@ -1,0 +1,1 @@
+export { GoalEditor } from "./GoalEditor";
