+++ src/features/permission-response/index.ts
@@ -1,0 +1,1 @@
+export { usePermissionResponse } from "./usePermissionResponse";
