+++ src/shared/ui/index.ts
@@ -1,0 +1,6 @@
+export { Badge, badgeVariants } from "./Badge";
+export { Button, buttonVariants } from "./Button";
+export { Card, CardContent, CardHeader, CardTitle, CardTitleBlock } from "./Card";
+export { Input } from "./Input";
+export { NativeSelect } from "./NativeSelect";
+export { Textarea } from "./Textarea";
