+++ src/widgets/event-stream/index.ts
@@ -1,0 +1,1 @@
+export { EventStream } from "./EventStream";
