+++ src/widgets/follow-up-composer/index.ts
@@ -1,0 +1,1 @@
+export { FollowUpComposer } from "./FollowUpComposer";
