+++ src/widgets/follow-up-queue/index.ts
@@ -1,0 +1,1 @@
+export { FollowUpQueue } from "./FollowUpQueue";
