+++ src/widgets/workbench-tabs/index.ts
@@ -1,0 +1,1 @@
+export { TabBar } from "./TabBar";
